Use modified version of arena allocator
The arena allocator defined in cmark-gfm is not thread safe. This can
cause issues when using Markable across multiple threads. This commit
replaces the version of arena.c included with cmark-gfm. The replaced
version uses the JANET_THREAD_LOCAL macro definition to make the value
of A local to each thread.